+++ src/widgets/spinbox.rs
@@ -215,7 +215,7 @@ impl<T: PartialOrd + AddAssign + SubAssign + ToString + Copy + Display + NumCast
         let window = ui.context.window.clone();
         std::thread::spawn(move || {
             std::thread::sleep(std::time::Duration::from_millis(st));
-            window.request_update(UserEvent::ReqUpdate);
+            window.request_update_event(UserEvent::ReqUpdate);
         });
         // #[cfg(feature = "winit")]
         // {