+++ src/window/mod.rs
@@ -82,6 +82,7 @@ pub enum UserEvent {
     ReqUpdate = 0,
     CreateChild = 1,
     ReInit = 2,
+    UserUpdate = 3,
 }
 
 pub struct WindowType {
@@ -94,31 +95,31 @@ pub struct WindowType {
 
 
 impl WindowType {
-    pub const ROOT: u32 = 0;
-    pub const CHILD: u32 = 1;
+    pub(crate) const ROOT: u32 = 0;
+    pub(crate) const CHILD: u32 = 1;
 
     #[cfg(all(target_os = "linux", not(feature = "winit")))]
-    pub fn x11(&self) -> &X11WindowHandle {
+    pub(crate) fn x11(&self) -> &X11WindowHandle {
         match self.kind {
             WindowKind::X11(ref window) => window,
         }
     }
 
     #[cfg(feature = "winit")]
-    pub fn winit(&self) -> &WInitWindowHandle {
+    pub(crate) fn winit(&self) -> &WInitWindowHandle {
         match self.kind {
             WindowKind::Winit(ref window) => window,
         }
     }
 
     #[cfg(all(not(feature = "winit"), target_os = "windows"))]
-    pub fn win32(&self) -> &win32::handle::Win32WindowHandle {
+    pub(crate) fn win32(&self) -> &win32::handle::Win32WindowHandle {
         match self.kind {
             WindowKind::Win32(ref window) => window
         }
     }
 
-    pub fn set_ime_position(&self, x: f32, y: f32, cursor_height: f32) {
+    pub(crate) fn set_ime_position(&self, x: f32, y: f32, cursor_height: f32) {
         match self.kind {
             #[cfg(all(target_os = "linux", not(feature = "winit")))]
             WindowKind::X11(ref window) => window.set_ime_position(&self.ime, x, y + cursor_height),
@@ -129,11 +130,11 @@ impl WindowType {
         }
     }
 
-    pub fn id(&self) -> WindowId {
+    pub(crate) fn id(&self) -> WindowId {
         self.id
     }
 
-    pub fn request_redraw(&self) {
+    pub(crate) fn request_redraw(&self) {
         match self.kind {
             #[cfg(all(target_os = "linux", not(feature = "winit")))]
             WindowKind::X11(ref window) => window.request_redraw(),
@@ -144,7 +145,7 @@ impl WindowType {
         }
     }
 
-    pub fn request_update(&self, event: UserEvent) {
+    pub(crate) fn request_update_event(&self, event: UserEvent) {
         match self.kind {
             #[cfg(all(target_os = "linux", not(feature = "winit")))]
             WindowKind::X11(ref window) => window.send_update(event),
@@ -155,19 +156,23 @@ impl WindowType {
         }
     }
 
-    pub fn ime(&self) -> &Arc<IME> {
+    pub fn request_update(&self) {
+        self.request_update_event(UserEvent::UserUpdate);
+    }
+
+    pub(crate) fn ime(&self) -> &Arc<IME> {
         &self.ime
     }
 
     #[cfg(all(not(feature = "winit"), target_os = "windows"))]
-    pub fn set_visible(&self, visible: bool) {
+    pub(crate) fn set_visible(&self, visible: bool) {
         match self.kind {
             WindowKind::Win32(ref window) => window.set_visible(visible).unwrap(),
         }
     }
 
     #[cfg(not(feature = "winit"))]
-    pub fn request_clipboard(&self, clipboard: ClipboardData) {
+    pub(crate) fn request_clipboard(&self, clipboard: ClipboardData) {
         match self.kind {
             #[cfg(all(target_os = "linux", not(feature = "winit")))]
             WindowKind::X11(ref window) => window.request_clipboard(clipboard),
@@ -177,7 +182,7 @@ impl WindowType {
     }
 
     #[cfg(not(feature = "winit"))]
-    pub fn set_clipboard(&self, clipboard: ClipboardData) {
+    pub(crate) fn set_clipboard(&self, clipboard: ClipboardData) {
         match self.kind {
             #[cfg(all(target_os = "linux", not(feature = "winit")))]
             WindowKind::X11(ref window) => window.set_clipboard(clipboard),